+++ net/message/version.go
@@ -25,7 +25,7 @@ type version struct {
 		Services     uint64
 		TimeStamp    uint32
 		Port         uint16
-		HttpInfoPort uint16
+		HttpInfoPort int
 		Cap          [32]byte
 		Nonce        uint64
 		// TODO remove tempory to get serilization function passed
